refactor: move `formatCpp` function as `formatValue` under `generateCppVariableDeclarationCode`
